Update Q7.ply
Siren wishes to create a program to help produce a banner for her Halloween party. 
The program should prompt the user to enter a message. The program should then prompt the user to choose one of options shown below.  Once the user has selected an option, the program should carry out the selected option.